+++ src/main/java/io/andrewtxt/sudoku/model/Table.java
@@ -11,7 +11,7 @@ public class Table {
 
     private final List<List<Cell>> cells;
 
-    public Table(int[][] values) {
+    public Table(Byte[][] values) {
         this.cells = toCells(values);
         connectCells();
     }
@@ -49,12 +49,12 @@ public class Table {
         return String.join("\n", lines);
     }
 
-    private List<List<Cell>> toCells(int[][] values) {
+    private List<List<Cell>> toCells(Byte[][] values) {
         List<List<Cell>> result = new ArrayList<>();
-        for (int i = 0; i < ROW_NUMBER; i++) {
+        for (byte i = 0; i < ROW_NUMBER; i++) {
             result.add(new ArrayList<>());
-            for (int j = 0; j < COLUMN_NUMBER; j++) {
-                result.get(i).add(new Cell(i, j, (byte) values[i][j], this));
+            for (byte j = 0; j < COLUMN_NUMBER; j++) {
+                result.get(i).add(new Cell(i, j, values[i][j], this));
             }
         }
         return result;
